+++ packages/core/test/bundling/defer/bundle.golden_symbols.json
@@ -1116,9 +1116,6 @@
     "name": "init_authoring"
   },
   {
-    "name": "init_base_dispatcher"
-  },
-  {
     "name": "init_bindings"
   },
   {
